requirements.txt
altgraph==0.17.2
annotated-types==0.7.0
appdirs==1.4.4
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
configparser==7.2.0
confluent-kafka==2.4.0
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka-python==2.0.2
kiwisolver==1.4.8
logging==0.4.9.6
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.1
nest-asyncio==1.6.0
netprotocols==0.5.0
networkx==3.4.2
numpy==2.2.5
orjson==3.10.10
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pip==24.0
psutil==7.0.0
puremagic==1.28
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2022.7
pyparsing==3.2.3
pypdf==5.4.0
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
pyshark==0.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
quixstreams==3.1.1
referencing==0.35.1
requests==2.32.3
rocksdict==0.3.23
rpds-py==0.20.0
scapy==2.6.1
scikit-fuzzy==0.5.0
scipy==1.15.2
setuptools==75.3.0
six==1.16.0
termcolor==2.5.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
